chore: refactor controllers to move db logic to each respective models
