+++ controllers/BookReqController.js
@@ -1,4 +1,4 @@
-const DbConn = require("../helper/DbTransaction");
+const BookReqModel = require("../models/BookReqModel");
 const Emailer = require("../helper/Emailer");
 const Notif = require("../helper/NotificationManager");
 const cron = require("node-cron");
@@ -14,33 +14,18 @@ const convertTZ = require("../helper/helper");
 //           ELSE ''
 //         END AS status
 
-const BookReqCol = [
-  "id_ruangan",
-  "id_user",
-  "created_at",
-  "book_date",
-  "time_start",
-  "time_end",
-  "agenda",
-  "prtcpt_ctr",
-  "remark",
-  "is_active",
-  "id_book",
-];
-
 const BookReqController = {
   createBook: async (req, res) => {
     const data = req.body.data;
 
     // Validate required fields
     if (!data.id_ruangan || !data.category || !data.participant) {
-      return res.status(400).send({
+      res.status(400).send({
         message: "Missing required fields: room ID, category, and participant count are required",
       });
+      return;
     }
 
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     let today = new Date();
     today = convertTZ(today, "Asia/Jakarta");
 
@@ -48,38 +33,27 @@ const BookReqController = {
     const id_notif = uuid.uuid();
 
     try {
-      await client.beginTransaction();
+      const bookingData = {
+        ...data,
+        created_at: today,
+      };
 
-      // ATOMIC AVAILABILITY CHECK - within the same transaction
-      const isBooked = await client.query(
-        `SELECT
-          id_ruangan,
-          DATE_FORMAT(book_date, '%Y-%m-%d') as book_date,
-          DATE_FORMAT(time_start, '%H:%i') as time_start,
-          DATE_FORMAT(time_end, '%H:%i') as time_end
-        FROM
-          req_book
-        WHERE
-          id_ruangan = ?
-          AND book_date = ?
-          AND is_active = 'T'
-          AND (
-            (req_book.time_start < ? AND req_book.time_end > ?)
-          )
-        FOR UPDATE`, // Add FOR UPDATE to lock the rows during transaction
-        [data.id_ruangan, data.book_date, data.time_end, data.time_start]
-      );
+      const result = await BookReqModel.createBooking(bookingData, id_book, id_notif);
 
-      // If room is already booked, send rejection email and fail
-      if (isBooked[0].length > 0) {
-        await client.rollback();
-
+      if (!result.success && result.conflict) {
         // Get user info for rejection email
-        const userInfo = await client.query("SELECT email, username FROM mst_user WHERE id_user = ?", [data.id_user]);
+        const userInfo = await BookReqModel.getUserEmail(data.id_user);
+
+        if (!userInfo) {
+          res.status(404).send({
+            message: "User not found",
+          });
+          return;
+        }
 
         const rejectionData = {
-          email: userInfo[0][0].email,
-          username: userInfo[0][0].username,
+          email: userInfo.email,
+          username: userInfo.username,
           approval: "rejected",
           reject_note: `The room ${data.id_ruangan} is already booked for the requested time slot. Please choose a different time or room.`,
           agenda: data.agenda,
@@ -95,43 +69,12 @@ const BookReqController = {
         const Email = new Emailer();
         await Email.approvalNotif(rejectionData);
 
-        return res.status(400).send({
+        res.status(400).send({
           message: `${data.id_ruangan} is already booked for this time slot`,
-          booked: isBooked[0],
+          booked: result.booked,
         });
+        return;
       }
-
-      // Room is available - create booking immediately as approved
-      const payload = {
-        id_ruangan: data.id_ruangan,
-        id_user: data.id_user,
-        created_at: today,
-        book_date: data.book_date,
-        time_start: data.time_start,
-        time_end: data.time_end,
-        agenda: data.agenda,
-        prtcpt_ctr: data.participant,
-        remark: data.remark,
-        category: data.category,
-        id_book: id_book,
-        is_active: "T",
-        id_notif: id_notif,
-        approval: "approved", // IMMEDIATELY APPROVED - no pending status
-        check_in: "F",
-        check_out: "F",
-      };
-
-      const [query, value] = await Client.insertQuery(payload, "req_book");
-      await client.query(query, value);
-
-      const n = await client.query("SELECT nama FROM mst_user WHERE id_user = ?", [payload.id_user]);
-      Object.defineProperty(payload, "nama", { value: n[0][0].nama });
-
-      const q = await client.query("SELECT id_ticket FROM req_book where id_book = ?", [id_book]);
-      const id_ticket = q[0][0].id_ticket;
-
-      // Commit the transaction first to ensure booking is saved
-      await client.commit();
 
       // Set up notifications since booking is immediately approved
       const bookDate = moment(new Date(`${data.book_date} ${data.time_start}`)).subtract(15, "m");
@@ -143,21 +86,28 @@ const BookReqController = {
         id_book,
         id_notif
       );
-      await Notif.CreateNewCronMail(bookDate, payload);
+      await Notif.CreateNewCronMail(bookDate, result.payload);
 
       // Send confirmation email to user (not admin notification)
-      const userEmail = await client.query("SELECT email, username FROM mst_user WHERE id_user = ?", [payload.id_user]);
+      const userEmail = await BookReqModel.getUserEmail(result.payload.id_user);
+
+      if (!userEmail) {
+        res.status(404).send({
+          message: "User not found",
+        });
+        return;
+      }
       const userData = {
-        email: userEmail[0][0].email,
-        username: userEmail[0][0].username,
+        email: userEmail.email,
+        username: userEmail.username,
         approval: "approved",
-        agenda: payload.agenda,
-        remark: payload.remark,
-        ruangan: payload.id_ruangan, // Map id_ruangan to ruangan for email template
-        book_date: payload.book_date,
-        time_start: payload.time_start,
-        time_end: payload.time_end,
-        capacity: payload.prtcpt_ctr, // Map prtcpt_ctr to capacity for email template
+        agenda: result.payload.agenda,
+        remark: result.payload.remark,
+        ruangan: result.payload.id_ruangan, // Map id_ruangan to ruangan for email template
+        book_date: result.payload.book_date,
+        time_start: result.payload.time_start,
+        time_end: result.payload.time_end,
+        capacity: result.payload.prtcpt_ctr, // Map prtcpt_ctr to capacity for email template
       };
 
       const Email = new Emailer();
@@ -165,84 +115,65 @@ const BookReqController = {
 
       res.status(200).send({
         message: "Room booked successfully! No approval needed.",
-        id_ticket: id_ticket,
+        id_ticket: result.id_ticket,
         status: "approved",
       });
     } catch (error) {
       console.log(error);
-      await client.rollback();
       res.status(500).send({ message: error.message });
-    } finally {
-      client.release();
     }
   },
 
   getBookById: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
-      await client.beginTransaction();
       const value = req.params.id_book;
-      const query =
-        "SELECT req_book.*, mst_user.username, mst_user.email, mst_room.nama as nama_ruangan, mst_room.kapasitas, mst_room.lokasi, mst_room.image, mst_room.is_virtual, mst_room.zoom_link, mst_room.zoom_meeting_id, mst_room.zoom_passcode FROM req_book LEFT JOIN mst_user ON req_book.id_user = mst_user.id_user LEFT JOIN mst_room ON req_book.id_ruangan = mst_room.id_ruangan WHERE id_book = ?";
-      const data = await client.query(query, value);
-      console.log(data, "getbookbyid");
-      await client.commit();
-      res.status(200).send(data[0][0]);
-      console.log(data);
+      const data = await BookReqModel.getBookById(value);
+
+      if (!data) {
+        res.status(404).send({
+          message: "Booking not found",
+        });
+        return;
+      }
+      res.status(200).send(data);
     } catch (error) {
-      await client.rollback();
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   editBook: async (req, res) => {
     let today = new Date();
     today = convertTZ(today, "Asia/Jakarta");
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const data = req.body.data;
-      console.log(data);
       const id_book = req.params.id_book;
       if (!id_book) {
         throw Error("Request Error");
       }
-      // Check availability for the new time slot (excluding current booking)
-      const isBooked = await client.query(
-        `SELECT
-          id_ruangan,
-          DATE_FORMAT(book_date, '%Y-%m-%d') as book_date,
-          DATE_FORMAT(time_start, '%H:%i') as time_start,
-          DATE_FORMAT(time_end, '%H:%i') as time_end
-        FROM
-          req_book
-        WHERE
-          id_ruangan = ?
-          AND book_date = ?
-          AND is_active = 'T'
-          AND id_book != ?
-          AND (
-            (req_book.time_start < ? AND req_book.time_end > ?)
-          )
-        FOR UPDATE`,
-        [data.id_ruangan, data.book_date, id_book, data.time_end, data.time_start]
-      );
 
-      // If new time slot is already booked, send rejection email and fail
-      if (isBooked[0].length > 0) {
-        await client.rollback();
+      const updateData = {
+        ...data,
+        updated_at: today,
+      };
 
+      const result = await BookReqModel.updateBooking(updateData, id_book);
+
+      if (!result.success && result.conflict) {
         // Get user info for rejection email
-        const userInfo = await client.query("SELECT email, username FROM mst_user WHERE id_user = ?", [data.id_user]);
+        const userInfo = await BookReqModel.getUserEmail(data.id_user);
+
+        if (!userInfo) {
+          res.status(404).send({
+            message: "User not found",
+          });
+          return;
+        }
 
         const rejectionData = {
-          email: userInfo[0][0].email,
-          username: userInfo[0][0].username,
+          email: userInfo.email,
+          username: userInfo.username,
           approval: "rejected",
           reject_note: `Cannot update booking: The room ${data.id_ruangan} is already booked for the new requested time slot. Please choose a different time or room.`,
           agenda: data.agenda,
@@ -258,45 +189,25 @@ const BookReqController = {
         const Email = new Emailer();
         await Email.approvalNotif(rejectionData);
 
-        return res.status(400).send({
+        res.status(400).send({
           message: `${data.id_ruangan} is already booked for this time slot`,
-          booked: isBooked[0],
+          booked: result.booked,
         });
+        return;
       }
 
-      const payload = {
-        id_ruangan: data.id_ruangan,
-        book_date: data.book_date,
-        time_start: data.time_start,
-        time_end: data.time_end,
-        agenda: data.agenda,
-        prtcpt_ctr: data.participant,
-        remark: data.remark,
-        approval: "approved", // Keep as approved since no manual approval needed
-        updated_at: today,
-        updated_by: data.id_user,
-      };
-      await client.beginTransaction();
-      console.log("BEFORE", cron.getTasks());
-      let id_notif = "";
-      const notif = await client.query(`SELECT id_notif FROM push_sched WHERE id_req = ? AND type = 'push'`, [id_book]);
-      if (notif[0][0]) {
-        id_notif = notif[0][0].id_notif;
-        await client.query(`DELETE FROM push_sched WHERE id_req = ?`, [id_book]);
-        global.scheduledTasks.delete(id_notif);
-      }
-      const [query, value] = Client.updateQuery(payload, { id_book: id_book }, "req_book");
-      const updateData = await client.query(query, value);
-      const q = await client.query("SELECT id_ticket from req_book where id_book = ?", [id_book]);
-      const id_ticket = q[0][0].id_ticket;
-      await client.commit();
+      const userInfo = await BookReqModel.getUserEmail(data.id_user);
 
-      const userInfo = await client.query("SELECT nama, email, username FROM mst_user WHERE id_user = ?", [
-        data.id_user,
-      ]);
+      if (!userInfo) {
+        res.status(404).send({
+          message: "User not found",
+        });
+        return;
+      }
+
       const userData = {
-        email: userInfo[0][0].email,
-        username: userInfo[0][0].username,
+        email: userInfo.email,
+        username: userInfo.username,
         approval: "approved",
         agenda: data.agenda,
         remark: data.remark,
@@ -312,82 +223,45 @@ const BookReqController = {
 
       res.status(200).send({
         message: "Booking updated successfully! No approval needed.",
-        id_ticket: id_ticket,
+        id_ticket: result.id_ticket,
         status: "approved",
       });
-      console.log(query, value);
-      console.log(updateData);
     } catch (error) {
-      await client.rollback();
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   cancelBook: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const id_book = req.params.id_book;
-      await client.beginTransaction();
-
-      const [query, value] = Client.updateQuery(
-        { is_active: "F", approval: "canceled" },
-        { id_book: id_book },
-        "req_book"
-      );
-      const updateData = await client.query(query, value);
-      await client.commit();
+      await BookReqModel.cancelBooking(id_book);
       res.status(200).send({
         message: `${id_book} is canceled`,
       });
     } catch (error) {
-      await client.rollback();
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   showAllBook: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
-      await client.beginTransaction();
       const book_date = req.query.book_date || null;
       const approval = req.query.approval || null;
       const room = req.query.room || null;
 
-      let approvalFilter = approval === "calendar" ? ["approved", "finished", "pending"] : [approval];
-      let approvalPlaceholders = approvalFilter.map(() => "?").join(",");
-
-      const showall = await client.query(
-        `SELECT req_book.*, mst_user.username FROM req_book LEFT JOIN mst_user ON req_book.id_user = mst_user.id_user
-        WHERE (req_book.book_date = ? OR ? IS NULL)
-        AND (req_book.approval IN (${approvalPlaceholders}) OR ? IS NULL)
-        AND (req_book.id_ruangan = ? OR ? IS NULL)
-        ORDER BY req_book.id DESC`,
-        [book_date, book_date, ...approvalFilter, approval, room, room]
-      );
-      await client.commit();
-      res.status(200).send({ data: showall[0] });
+      const data = await BookReqModel.getAllBookings(book_date, approval, room);
+      res.status(200).send({ data });
     } catch (error) {
-      await client.rollback();
       console.error(error);
       res.status(500).send(error);
-    } finally {
-      client.release();
     }
   },
 
   showBookbyUser: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const userid = req.query.id_user;
       const book_date = req.query.book_date || null;
@@ -397,97 +271,34 @@ const BookReqController = {
       if (userid === undefined) {
         throw Error("Request Error");
       }
-      await client.beginTransaction();
-      const showData = await client.query(
-        `SELECT
-        id_book,
-        id_ticket,
-        id_user,
-        MR.nama as nama_ruangan,
-        MR.id_ruangan as id_room,
-        MR.is_virtual,
-        MR.zoom_link,
-        MR.zoom_meeting_id,
-        MR.zoom_passcode,
-        agenda,
-        BK.is_active,
-        BK.approval,
-        time_start,
-        time_end,
-        book_date,
-        approval
-      FROM
-        (
-        SELECT
-          id_book,
-          id_ticket,
-          id_ruangan,
-          agenda,
-          id_user,
-          is_active,
-          approval,
-          DATE_FORMAT(time_start, '%H:%i') as time_start,
-          DATE_FORMAT(book_date, '%Y-%m-%d') as book_date,
-          DATE_FORMAT(time_end, '%H:%i') as time_end,
-          TIMESTAMP (
-          CONCAT( book_date, ' ', time_start )) AS start_time,
-          TIMESTAMP (
-          CONCAT( book_date, ' ', time_end)) AS end_time,
-          TIMESTAMP ( TIMESTAMP (
-          CONCAT( book_date, ' ', time_start )) - INTERVAL 15 MINUTE ) AS upcoming_time
-        FROM
-        req_book
-        ) BK
-        LEFT JOIN mst_room MR ON BK.id_ruangan = MR.id_ruangan
-        WHERE id_user = ?
-          AND
-            (book_date = ? OR ? IS NULL)
-          AND
-            (BK.is_active = ? OR ? IS NULL)
-          AND
-            (BK.approval = ? OR ? IS NULL)
-        ORDER BY book_date DESC
-        LIMIT ?`,
-        [userid, book_date, book_date, active, active, status, status, limit]
-      );
-      const data = showData[0];
-      await client.commit();
-      res.status(200).send({ data: data });
+
+      const data = await BookReqModel.getBookingsByUser(userid, book_date, limit, status, active);
+      res.status(200).send({ data });
     } catch (error) {
-      await client.rollback();
       console.error(error);
       if (error.message === "Request Error") {
         res.status(400).send({ message: error.message });
       } else {
         res.status(500).send({ message: error.message });
       }
-    } finally {
-      client.release();
     }
   },
 
   showBookbyRoom: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     const roomId = req.query.roomid;
     try {
       if (roomId === undefined) {
         throw Error("Request Error");
       }
-      const get = await client.query("SELECT * FROM req_book where id_ruangan = ? and is_active = 'F'", [roomId]);
-      const books = get[0];
+      const books = await BookReqModel.getBookingsByRoom(roomId);
       res.status(200).send({ data: books });
     } catch (error) {
       console.log(error);
       res.status(500).send(error);
-    } finally {
-      client.release();
     }
   },
 
   approval: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const data = req.body.data;
       const id_book = req.params.id_book;
@@ -501,10 +312,8 @@ const BookReqController = {
         reject_note: data.reject_note,
       };
       console.log(payload);
-      await client.beginTransaction();
-      const [query, value] = Client.updateQuery(payload, { id_book: id_book }, "req_book");
-      const updateData = await client.query(query, value);
-      await client.commit();
+
+      await BookReqModel.updateApproval(payload, id_book);
 
       if (data.approval === "approved") {
         await Notif.CreateNewCron(
@@ -525,265 +334,113 @@ const BookReqController = {
         reject_note: data.reject_note,
       });
     } catch (error) {
-      await client.rollback();
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   checkIn: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const { id_user, room_id } = req.body.data;
       console.log({ id_user, room_id }, "check-in data");
 
       if (!id_user || !room_id) {
-        return res.status(400).send({
+        res.status(400).send({
           message: "User ID and Room ID are required",
         });
+        return;
       }
 
-      await client.beginTransaction();
+      const result = await BookReqModel.checkIn(id_user, room_id);
 
-      // Find active booking for this user in this room today within check-in window
-      const bookingQuery = await client.query(
-        `SELECT * FROM req_book WHERE
-          id_user = ?
-          AND id_ruangan = ?
-          AND is_active = 'T'
-          AND check_in = 'F'
-          AND approval = 'approved'
-          AND book_date = DATE_FORMAT(CONVERT_TZ(NOW(), '+00:00', '+07:00'), '%Y-%m-%d')
-          AND CONVERT_TZ(CURTIME(), '+00:00', '+07:00') BETWEEN (time_start - INTERVAL 15 MINUTE) AND (time_start + INTERVAL 15 MINUTE)`,
-        [id_user, room_id]
-      );
-
-      if (bookingQuery[0].length === 0) {
-        await client.rollback();
-        return res.status(404).send({
-          message:
-            "No valid booking found for check-in. Make sure you have an active booking for this room today within the check-in window (15 minutes before start time).",
+      if (!result.success) {
+        res.status(404).send({
+          message: result.message,
         });
+        return;
       }
 
-      const booking = bookingQuery[0][0];
-      const id_book = booking.id_book;
-
-      // Update booking to checked in
-      const payload = { check_in: "T" };
-      const [query, value] = Client.updateQuery(payload, { id_book: id_book }, "req_book");
-
-      const updateData = await client.query(query, value);
-      await client.commit();
-
-      console.log(query, value);
-      console.log("Rows affected:", updateData[0].changedRows);
-
-      if (updateData[0].changedRows === 1) {
-        res.status(200).send({
-          message: "Check in successful",
-          id_user: id_user,
-          id_book: id_book,
-          room_id: room_id,
-          agenda: booking.agenda,
-          time_start: booking.time_start,
-          time_end: booking.time_end,
-        });
-      } else {
-        res.status(400).send({
-          message: "Failed to update check-in status",
-        });
-      }
+      res.status(200).send({
+        message: "Check in successful",
+        id_user: id_user,
+        id_book: result.booking.id_book,
+        room_id: room_id,
+        agenda: result.booking.agenda,
+        time_start: result.booking.time_start,
+        time_end: result.booking.time_end,
+      });
     } catch (error) {
-      await client.rollback();
       console.error("Check-in error:", error);
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   checkOut: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const { id_user, room_id } = req.body.data;
       console.log({ id_user, room_id }, "check-out data");
 
       if (!id_user || !room_id) {
-        return res.status(400).send({
+        res.status(400).send({
           message: "User ID and Room ID are required",
         });
+        return;
       }
 
-      await client.beginTransaction();
+      const result = await BookReqModel.checkOut(id_user, room_id);
 
-      // Find active booking for this user in this room today that's checked in but not checked out
-      const bookingQuery = await client.query(
-        `SELECT * FROM req_book WHERE
-          id_user = ?
-          AND id_ruangan = ?
-          AND is_active = 'T'
-          AND check_in = 'T'
-          AND check_out = 'F'
-          AND approval = 'approved'
-        `,
-        [id_user, room_id]
-      );
-
-      if (bookingQuery[0].length === 0) {
-        await client.rollback();
-        return res.status(404).send({
-          message:
-            "No valid booking found for check-out. Make sure you have checked in and the meeting is still active (within 15 minutes after end time).",
+      if (!result.success) {
+        res.status(404).send({
+          message: result.message,
         });
+        return;
       }
 
-      const booking = bookingQuery[0][0];
-      const id_book = booking.id_book;
-
-      // Update booking to checked out and finished
-      const payload = {
-        check_out: "T",
-        is_active: "F",
-        approval: "finished",
-      };
-
-      const [query, value] = Client.updateQuery(payload, { id_book: id_book }, "req_book");
-
-      const updateData = await client.query(query, value);
-      await client.commit();
-
-      console.log(query, value);
-      console.log("Rows affected:", updateData[0].changedRows);
-
-      if (updateData[0].changedRows === 1) {
-        res.status(200).send({
-          message: "Check out successful",
-          id_user: id_user,
-          id_book: id_book,
-          room_id: room_id,
-          agenda: booking.agenda,
-          time_start: booking.time_start,
-          time_end: booking.time_end,
-        });
-      } else {
-        res.status(400).send({
-          message: "Failed to update check-out status",
-        });
-      }
+      res.status(200).send({
+        message: "Check out successful",
+        id_user: id_user,
+        id_book: result.booking.id_book,
+        room_id: room_id,
+        agenda: result.booking.agenda,
+        time_start: result.booking.time_start,
+        time_end: result.booking.time_end,
+      });
     } catch (error) {
-      await client.rollback();
       console.error("Check-out error:", error);
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   getCheckInBook: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     const id_user = req.params.id_user;
     try {
-      await client.beginTransaction();
-      const getBook = await client.query(
-        `
-        SELECT
-          req_book.*,
-          mst_room.is_virtual,
-          mst_room.zoom_link,
-          mst_room.zoom_meeting_id,
-          mst_room.zoom_passcode
-        FROM req_book
-        LEFT JOIN mst_room ON req_book.id_ruangan = mst_room.id_ruangan
-        WHERE (
-          req_book.id_user = ?
-          AND
-          req_book.is_active = 'T'
-          AND
-          req_book.check_in = 'F'
-          AND
-          req_book.approval = 'approved'
-          AND
-          req_book.book_date = DATE_FORMAT(CONVERT_TZ(NOW(), '+00:00', '+07:00'), '%Y-%m-%d')
-          AND
-          CONVERT_TZ(CURTIME(), '+00:00', '+07:00') BETWEEN (req_book.time_start - INTERVAL 15 MINUTE) AND (req_book.time_start + INTERVAL 15 MINUTE)
-        )
-        `,
-        [id_user]
-      );
-      await client.commit();
-      res.status(200).send({ data: getBook[0] });
+      const data = await BookReqModel.getCheckInBookings(id_user);
+      res.status(200).send({ data });
     } catch (error) {
-      await client.rollback();
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   getCheckOutBook: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     const id_user = req.params.id_user;
     try {
-      await client.beginTransaction();
-      const getBook = await client.query(
-        `
-        SELECT
-          req_book.*,
-          mst_room.is_virtual,
-          mst_room.zoom_link,
-          mst_room.zoom_meeting_id,
-          mst_room.zoom_passcode
-        FROM req_book
-        LEFT JOIN mst_room ON req_book.id_ruangan = mst_room.id_ruangan
-        WHERE (
-          req_book.id_user = ?
-          AND
-          req_book.is_active = 'T'
-          AND
-          req_book.check_in = 'T'
-          AND
-          req_book.check_out = 'F'
-          AND
-          req_book.approval = 'approved'
-          AND
-          req_book.book_date = DATE_FORMAT(CONVERT_TZ(NOW(), '+00:00', '+07:00'), '%Y-%m-%d')
-          AND
-          CONVERT_TZ(CURTIME(), '+00:00', '+07:00') > req_book.time_start
-          AND
-          CONVERT_TZ(CURTIME(), '+00:00', '+07:00') < req_book.time_end + INTERVAL 15 MINUTE
-        );
-        `,
-        [id_user]
-      );
-      await client.commit();
-      res.status(200).send({ data: getBook[0] });
+      const data = await BookReqModel.getCheckOutBookings(id_user);
+      res.status(200).send({ data });
     } catch (error) {
-      await client.rollback();
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 
   cancelBooking: async (req, res) => {
-    const Client = new DbConn();
-    const client = await Client.initConnection();
     try {
       const id_book = req.params.id_book;
       const { cancel_reason } = req.body;
@@ -792,58 +449,28 @@ const BookReqController = {
         throw new Error("Book ID is required");
       }
 
-      await client.beginTransaction();
+      const result = await BookReqModel.cancelBookingWithReason(id_book, cancel_reason);
 
-      // Get booking details first
-      const bookingQuery = await client.query(
-        "SELECT req_book.*, mst_user.username, mst_user.email FROM req_book LEFT JOIN mst_user ON req_book.id_user = mst_user.id_user WHERE id_book = ?",
-        [id_book]
-      );
-
-      if (bookingQuery[0].length === 0) {
-        await client.rollback();
-        res.status(404).send({
-          message: "Booking not found",
+      if (!result.success) {
+        res.status(result.message === "Booking not found" ? 404 : 400).send({
+          message: result.message,
         });
         return;
       }
-
-      const booking = bookingQuery[0][0];
-
-      // Check if booking can be cancelled (only active bookings)
-      if (booking.is_active !== "T") {
-        await client.rollback();
-        res.status(400).send({
-          message: "Cannot cancel inactive booking",
-        });
-        return;
-      }
-
-      // Update booking to cancelled status
-      const payload = {
-        is_active: "F",
-        approval: "cancelled",
-        reject_note: cancel_reason || "Cancelled by admin",
-      };
-
-      const [query, value] = Client.updateQuery(payload, { id_book: id_book }, "req_book");
-
-      const updateData = await client.query(query, value);
-      await client.commit();
 
       // Send cancellation email to user
       const emailData = {
-        email: booking.email,
-        username: booking.username,
+        email: result.booking.email,
+        username: result.booking.username,
         approval: "cancelled",
         reject_note: cancel_reason || "Your booking has been cancelled by admin",
-        agenda: booking.agenda,
-        remark: booking.remark,
-        ruangan: booking.id_ruangan,
-        book_date: booking.book_date,
-        time_start: booking.time_start,
-        time_end: booking.time_end,
-        capacity: booking.prtcpt_ctr,
+        agenda: result.booking.agenda,
+        remark: result.booking.remark,
+        ruangan: result.booking.id_ruangan,
+        book_date: result.booking.book_date,
+        time_start: result.booking.time_start,
+        time_end: result.booking.time_end,
+        capacity: result.booking.prtcpt_ctr,
       };
 
       console.log(emailData, "emailData");
@@ -857,13 +484,10 @@ const BookReqController = {
         cancel_reason: cancel_reason,
       });
     } catch (error) {
-      await client.rollback();
       console.error(error);
       res.status(500).send({
         message: error.message,
       });
-    } finally {
-      client.release();
     }
   },
 };
